feat: avancée sur la box Likes Price
